Update message test
cli_alert_inform() and each element in cli_dl() return individual messages. These are not captured as a whole by expect_message(), but need to be tested individually (ignoring the rest with suppressMessages). We now test the first 3 rows.

The alternative (used in print()) is to use cli::cat_line(), which returns a single message. It is hard to mimic the entire styling of cli_alert_inform though, as one needs to define bullet and bullet colour etc.